refactor: move test helpers into a subfolder
